Move all endpoints to router module
Splits them from the app object
