+++ tests/unit/app/test_sentry.py
@@ -16,7 +16,7 @@ def test_errors_are_reported_to_sentry(
     anon_client, webhook_create_example: BugzillaWebhookRequest
 ):
     with patch("sentry_sdk.hub.Hub.capture_event") as mocked:
-        with patch("src.app.api.execute_action", side_effect=ValueError):
+        with patch("src.app.router.execute_action", side_effect=ValueError):
             with pytest.raises(ValueError):
                 anon_client.post(
                     "/bugzilla_webhook", data=webhook_create_example.json()
